Filter search result and keep only and keep only the first match

Same as tw_filter(), but consistently returns data frames with a single row.

tw_filter_first(
  search,
  p,
  q,
  language = "en",
  limit = 10,
  wait = 0,
  cache = NULL
)

Arguments

search
A data frame generated by tw_search(), or a search query. If a data frame is given, language and limits are ignore.

p
A character vector of length 1, a property. Must always start with the capital letter "P", e.g. "P31" for "instance of".

q
A character vector of length 1, a wikidata id. Must always start with the capital letter "Q", e.g. "Q5" for "human being".

language
Language to be used for the search.

limit
Maximum numbers of responses to be given.

wait
In seconds, defaults to 0. Time to wait between queries to Wikidata. If data are cached locally, wait time is not applied. If you are running many queries systematically you may want to add some waiting time between queries.

cache
Defaults to NULL. If given, it should be given either TRUE or FALSE. Typically set with tw_enable_cache() or tw_disable_cache()

Value

A data frame with one row and three columns, id, label, and description, filtered by the above criteria.

Examples

  tw_search("Margaret Mead") %>%
    tw_filter_first(p = "P31", q = "Q5")

#> # A tibble: 1 x 3
#>   id      label         description            
#>   <chr>   <chr>         <chr>                  
#> 1 Q180099 Margaret Mead American anthropologist