Filter search result and keep only people

A wrapper of tw_filter() that defaults to keep only "instance of" (P31) "human being" (Q5).

tw_filter_people(
  search,
  language = "en",
  limit = 10,
  stop_at_first = TRUE,
  wait = 0
)

Arguments

search

A data frame generated by tw_search(), or a search query. If a data frame is given, language and limits are ignore.

language

Language to be used for the search.

limit

Maximum numbers of responses to be given.

stop_at_first

Logical, defaults to TRUE. If TRUE, returns only the first match from the search that satisfies the criteria.

wait

In seconds, defaults to 0. Time to wait between queries to Wikidata. If data are cached locally, wait time is not applied. If you are running many queries systematically you may want to add some waiting time between queries.

Value

A data frame with three columns, id, label, and description; all rows refer to a human being.

Examples

tw_search("Ruth Benedict")
#> # A tibble: 4 × 3
#>   id        label                                    description                
#>   <chr>     <chr>                                    <chr>                      
#> 1 Q228822   Ruth Benedict                            American anthropologist an…
#> 2 Q80896797 Ruth Benedict                            scientific article publish…
#> 3 Q28002866 Ruth Benedict Prize                      LGBT anthropological award 
#> 4 Q58589016 Ruth Benedict, Boasian Anthropology, an… NA                         

tw_search("Ruth Benedict") %>%
  tw_filter_people()
#> # A tibble: 1 × 3
#>   id      label         description                               
#>   <chr>   <chr>         <chr>                                     
#> 1 Q228822 Ruth Benedict American anthropologist and folklorologist