Get Wikidata label in given language

Get Wikidata label in given language

tw_get_label(
  id,
  language = tidywikidatar::tw_get_language(),
  id_df = NULL,
  cache = NULL,
  overwrite_cache = FALSE,
  cache_connection = NULL,
  disconnect_db = TRUE,
  wait = 0
)

Arguments

id
A characther vector, must start with Q, e.g. "Q254" for Wolfgang Amadeus Mozart

language
Defaults to language set with tw_set_language(); if not set, "en". Use "all_available" to keep all languages. For available language values, see https://www.wikidata.org/wiki/Help:Wikimedia_language_codes/lists/all

id_df
Default to NULL. If given, it should be a dataframe typically generated with tw_get_(), and is used instead of calling Wikidata or using SQLite cache. Ignored when id is of length more than one.

cache
Defaults to NULL. If given, it should be given either TRUE or FALSE. Typically set with tw_enable_cache() or tw_disable_cache().

overwrite_cache
Logical, defaults to FALSE. If TRUE, it overwrites the table in the local sqlite database. Useful if the original Wikidata object has been updated.

cache_connection
Defaults to NULL. If NULL, and caching is enabled, tidywikidatar will use a local sqlite database. A custom connection to other databases can be given (see vignette caching for details).

disconnect_db
Defaults to TRUE. If FALSE, leaves the connection to cache open.

wait
In seconds, defaults to 0. Time to wait between queries to Wikidata. If data are cached locally, wait time is not applied. If you are running many queries systematically you may want to add some waiting time between queries.

Value

A charachter vector of the same length as the vector of id given, with the Wikidata label in the requested languae.

Examples

tw_get_label(
  id = c(
    "Q180099",
    "Q228822"
  ),
  language = "en"
)

#> [1] "Margaret Mead" "Ruth Benedict"

# If a label is not available, a NA value is returned
tw_get_label(
  id = c(
    "Q64733534",
    "Q4773904",
    "Q220480"
  ),
  language = "sc"
)

#> [1] NA                   NA                   "Zora Neale Hurston"