Search for Wikidata properties in Wikidata and return Wikidata id, label, and description.

This search returns only items, use tw_search_property() for properties.

tw_search_item(
  search,
  language = tidywikidatar::tw_get_language(),
  limit = 10,
  include_search = FALSE,
  wait = 0,
  cache = NULL,
  overwrite_cache = FALSE,
  cache_connection = cache_connection
)

Arguments

search
A string to be searched in Wikidata

language
Language to be used for the search. Can be set once per session with tw_set_language(). If not set, defaults to "en". For a full list, see https://www.wikidata.org/wiki/Help:Wikimedia_language_codes/lists/all

limit
Maximum numbers of responses to be given.

include_search
Logical, defaults to FALSE. If TRUE, the search is returned as an additional column.

wait
In seconds, defaults to 0.1. Time to wait between queries to Wikidata. If data are cached locally, wait time is not applied.

cache
Defaults to NULL. If given, it should be given either TRUE or FALSE. Typically set with tw_enable_cache() or tw_disable_cache().

overwrite_cache
Defaults to FALSE. If TRUE, overwrites cache.

cache_connection
Defaults to NULL. If NULL, and caching is enabled, tidywikidatar will use a local sqlite database. A custom connection to other databases can be given (see vignette caching for details).

Value

A data frame (a tibble) with three columns (id, label, and description), and as many rows as there are results (by default, limited to 10).

Examples

tw_search_item(search = "Sylvia Pankhurst")

#> # A tibble: 2 × 3
#>   id        label            description                                     
#>   <chr>     <chr>            <chr>                                           
#> 1 Q298213   Sylvia Pankhurst Activist, suffragette, pacifist internationalist
#> 2 Q24298894 Sylvia Pankhurst sculpture in Mile End Park, London