Return (most) information from a Wikidata item in a tidy format

Return (most) information from a Wikidata item in a tidy format

tw_get(
  id,
  language = "all_available",
  cache = NULL,
  overwrite_cache = FALSE,
  wait = 0,
  include_id = TRUE
)

Arguments

id
A characther vector, must start with Q, e.g. "Q180099" for the anthropologist Margaret Mead. Can also be a data frame of one row, typically generated with tw_search() or a combination of tw_search() and tw_filter_first().

language
Defaults to "all_available". By default, returns dataset with labels in all available languages. If given, only in the chosen language. For available values, see https://www.wikidata.org/wiki/Help:Wikimedia_language_codes/lists/all

cache
Defaults to NULL. If given, it should be given either TRUE or FALSE. Typically set with tw_enable_cache() or tw_disable_cache().

overwrite_cache
Logical, defaults to FALSE. If TRUE, it overwrites the table in the local sqlite database. Useful if the original Wikidata object has been updated.

wait
In seconds, defaults to 0. Time to wait between queries to Wikidata. If data are cached locally, wait time is not applied. If you are running many queries systematically you may want to add some waiting time between queries.

include_id
Logical, defaults to TRUE If TRUE, output includes a column with the wikidata id of the item.

Value

A data.frame (a tibble) a tibble with three columns, if include_id is set to TRUE (default). A tibble with two columns (property and value) if include_id is set to FALSE.

Examples

tw_get(
  id = "Q180099",
  language = "en"
)

#> # A tibble: 232 x 3
#>    id      property value        
#>    <chr>   <chr>    <chr>        
#>  1 Q180099 label_en Margaret Mead
#>  2 Q180099 P21      Q6581072     
#>  3 Q180099 P214     44302511     
#>  4 Q180099 P106     Q674426      
#>  5 Q180099 P106     Q4773904     
#>  6 Q180099 P106     Q36180       
#>  7 Q180099 P106     Q2526255     
#>  8 Q180099 P373     Margaret Mead
#>  9 Q180099 P244     n78093416    
#> 10 Q180099 P227     118579789    
#> # … with 222 more rows