Search for Wikidata items or properties and return Wikidata id, label, and description.

By defaults, this search returns items. Set type to property or use tw_search_property() for properties.

tw_search(
  search,
  type = "item",
  language = tidywikidatar::tw_get_language(),
  limit = 10,
  include_search = FALSE,
  wait = 0,
  cache = NULL,
  overwrite_cache = FALSE,
  cache_connection = NULL,
  disconnect_db = TRUE
)

Arguments

search
A string to be searched in Wikidata

type
Defaults to "item". Either "item" or "property".

language
Language to be used for the search. Can be set once per session with tw_set_language(). If not set, defaults to "en". For a full list, see https://www.wikidata.org/wiki/Help:Wikimedia_language_codes/lists/all

limit
Maximum numbers of responses to be given.

include_search
Logical, defaults to FALSE. If TRUE, the search is returned as an additional column.

wait
In seconds, defaults to 0. Time to wait between queries to Wikidata. If data are cached locally, wait time is not applied. If you are running many queries systematically you may want to add some waiting time between queries.

cache
Defaults to NULL. If given, it should be given either TRUE or FALSE. Typically set with tw_enable_cache() or tw_disable_cache().

overwrite_cache
Defaults to FALSE. If TRUE, overwrites cache.

cache_connection
Defaults to NULL. If NULL, and caching is enabled, tidywikidatar will use a local sqlite database. A custom connection to other databases can be given (see vignette caching for details).

disconnect_db
Defaults to TRUE. If FALSE, leaves the connection to cache open.

Value

A data frame (a tibble) with three columns (id, label, and description), and as many rows as there are results (by default, limited to 10). Four columns when include_search is set to TRUE.

Examples

tw_search(search = c("Margaret Mead", "Ruth Benedict"))

#> Warning: Unknown or uninitialised column: `search`.

#> # A tibble: 14 × 4
#>    id        label                              value description               
#>    <chr>     <chr>                              <chr> <chr>                     
#>  1 Q180099   Margaret Mead                      NA    American anthropologist   
#>  2 Q81015029 Margaret mead                      NA    scientific article publis…
#>  3 Q66701460 Margaret Mead                      NA    scientific article publis…
#>  4 Q85724626 Mead & Bateson                     NA    business organisation     
#>  5 Q96077616 Margaret Meadows                   NA    1718-1781                 
#>  6 Q76238541 Margaret Meadowe                   NA    Peerage person ID=628312  
#>  7 Q75506638 Margaret Meadows                   NA    Peerage person ID=183057  
#>  8 Q75812372 Margaret Meade-Waldo               NA    died 1954                 
#>  9 Q6759717  Margaret Mead Film Festival        NA    annual film festival held…
#> 10 Q55897055 Margaret Mead and Samoa: Coming o… NA    NA                        
#> 11 Q228822   Ruth Benedict                      NA    American anthropologist a…
#> 12 Q80896797 Ruth Benedict                      NA    scientific article publis…
#> 13 Q28002866 Ruth Benedict Prize                NA    LGBT anthropological award
#> 14 Q58589016 Ruth Benedict, Boasian Anthropolo… NA    NA